Refactor binding code to split it up more.
Drop unused convertors.
Bring detector config into the main binding code.
