+++ python_bindings/src/include/detectors.hpp
